pythonProject/requirements.txt
requests~=2.31.0
pillow~=10.3.0